Convert to use webserver to get auth
This now runs a webserver which captures the auth redirect from Spotify. This means the user doesn't need to copy/paste the url. However, it does make the browser window popup, and this could be handled a lot better.

There's a small amount of error handling so that it handles spotify not playing music more gracefully.