Trying some basic re-auth
Mutex works, but there's a question as to how useful it is or not to do here when the message passing is fine enough.
If we put the update the currently playing in it's own thread, we will need some way to make sure the main loop thread doesn't
start until the webserver is loaded and, ideally, the first authentication is done. Unless we hold some state
in the SpotifyClient struct, which increases the complexity a bit.
Either way, this has seemingly broken things, but it might be due to a poor internet connection
on the train...