cache the spotify token and use where needed
Cache the token obtained by Spotify. When we try authenticating with Spotify
check to see if this token exists and is valid, and then use it. Otherwise
do the normal authentication process.

This also uses auto-refresh with each call, so that makes the refreshing
code obsolete, so that's been removed.